added FrontEnd UI

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 const asyncHandler= require('express-async-handler')
 const Goal = require('../model/goalModel')
+const User = require('../model/userModel')
 
 //@desc Get goals
 //@route GET /api/goals
 //@access Private
 const getGoals= asyncHandler(async (req,res)=>{
-    const goals = await Goal.find()
+    const goals = await Goal.find({user:req.user.id})
     res.status(200).json(goals)
 })
 
@@ -16,10 +17,11 @@ const setGoal= asyncHandler(async(req,res)=>{
     if(!req.body.text){
         res.status(400)
         throw new Error('Please add a text field')
-    };
+    }
 
     const goal=await Goal.create({
-        text:req.body.text
+        text:req.body.text,
+        user:req.user.id
     })
 
     res.status(200).json(goal)
@@ -36,7 +38,17 @@ const updateGoal= asyncHandler(async (req,res)=>{
         res.status(400)
         throw new Error('Goal not found')
     }
+    const user=await  User.findById(req.user.id)
+    if(!user){
+        res.status(400)
+        throw new Error('User not found')
 
+    }
+    //match logged in user with goal user
+    if(goal.user.toString()!==user.id){
+        res.status(400)
+        throw new Error('User not authorized')
+    }
     const updateGoal=await Goal.findByIdAndUpdate(req.params.id,req.body,{
         new:true,
     })
@@ -53,6 +65,17 @@ const deleteGoal= asyncHandler(async (req,res)=>{
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+    const user=await  User.findById(req.user.id)
+    if(!user){
+        res.status(400)
+        throw new Error('User not found')
+
+    }
+    //match logged in user with goal user
+    if(goal.user.toString()!==user.id){
+        res.status(400)
+        throw new Error('User not authorized')
     }
     await goal.deleteOne()
